+++ Changelog.md
@@ -7,6 +7,7 @@
 - Change the default model type from `tspan` to `image` to reduce chance of error for models without metadata
 - Fix audio clipping when using segments
 - Allow users to customize video-compare settings in the Settings menu
+- Force kill trtexec and vspipe processes when beginning workflow processing
 
 ## 0.10.2
 - Implement segment selection. Users can now select a small segment of a video to process and preview!